refactor(Locadora): refatorar total

+++ src/modulo-04-c-sharp/dia-03/Locadora/Locadora.UI/Operacao.cs
@@ -22,7 +22,7 @@ namespace Locadora.UI
 
         // Variaveis para utilização nos métodos
         string nome;
-        double preco;
+        decimal preco;
         Categoria categoria;
 
         // Método para selecionar Operação a ser executada
